Userland mouse daemon for new PCVT mouse features. The daemon must be started for the mouse to work. A common usage for a PS/2 mouse might be:
/usr/sbin/moused -p /dev/psm0 -M2=3 -m IntelliMouse

This sets the mouse port to /dev/psm0, maps the right mouse button to paste
(by default, the right mouse button is for "extending" selections and the
middle button pastes), and the -m flag sets the specific model, in this example
for a Microsoft IntelliMouse, which enables the wheel scrollback functionality.
Note that for wheel mice you also need "option INTELLIMOUSE" compiled into the
kernel.

This daemon comes from FreeBSD and was ported to OpenBSD by Jean-Baptiste
Marchand.