Remove an over-optimization from the polling loop of siop_scsicmd(), i.e. calling siop_print_info() immediately for async devices.
With the change to use xs->sc_link as parameter to siop_print_info()
this immediate call wouldn't work as the sc_link is not initialized
until AFTER the return from processing the INQUIRY command.

So, just use the default negotiation logic which is triggered by the
next command to be handled.