naddy@ pointed out that in the two situations
case WORD

and

var=WORD

Bourne and related/descendant shells such as our sh do *not* do
field splitting or globbing on WORD. Thus we can eliminate a large
number of unnecessary '"'s in case statements and variable
assignments.

This is a lightly adapted version of naddy's diff, to take account
of changes in the scripts since he generated it.