for every sensor esm looks after it has to talk to the hardware, which means it could busy wait while the hardware gets ready. all the sensors are updated out of a single timeout. doing this for the 30ish esm sensors on my machine is probably not a good thing...
this changes the way updates are handled. instead of a single big update
for all the sensors, we now update only one sensor in a timeout. after
that sensor has been updated it sets up a very short timeout for the next
sensor to run out of. we also use the short timeouts to retry the same
sensor if the hardware isnt ready which means we avoid the (worst) busy
wait.