Add code to change video mode (resolution and color depth) on the fly for pnozz(4), when switching between emulation (i.e. text console) and mapped (i.e.  X11) modes.
Geometries different than 800x600 are only available on the external video
port, with the internal panel blanked.

Currently this mode is compiled in, until an interface allows the X server
and the kernel to settle on which one to use.

Due to the internal panel blanking requirements, all of this is only available
if tctrl(4) is configured in, which is the case for GENERIC but not for
installation media kernels (who runs X11 with installation media anyway?)

Most of this has been written 18 months ago, it was just lacking a final
touch...