The default address family to be used is stored in a global variable.  If no options are given on the commandline, it is set to PF_INET.
The configuration file parser did not use this variable in all cases, but
used PF_UNSPEC for getaddrinfo/getnameinfo, leading to bogus error messages
in some cases (but httpd operated as expected).  Use the global variable
instead of the hardcode PF_UNSPEC in the cases.

Add a new commandline flag, -U, to set the default address family to
PF_UNSPEC for ambigous directives.

Discussed with sthen.