only look up the initiator-id for SPI (aka "real scsi") controllers on sparc64.
this prevents us from setting the initiator id on sas and fc hbas. the
scsi midlayer doesnt probe for targets at the initiator id address, so if
we set this on fc and sas hbas (which dont really have a real scsi address
on the bus), then we're preventing them from finding potential targets at
that address.

ok kettenis@